classify_habitats

Classify habitats

classify_habitats(raster, classes = 5, style = "fisher")

Arguments

raster
RasterLayer with continuous environmental values.

classes
Integer with number of classes.

style
Character with method of classification. See Details section for more information.

Value

RasterLayer

Details

Classifies a RasterLayer from the raster packages with continuous values into n discrete classes. Consequently, classes are non-overlapping (and left-closed). For more information about the classification method, see classIntervals from the classInt package and/or the provided References.

References

Armstrong, M. P., Xiao, N., Bennett, D. A., 2003. "Using genetic algorithms to create multicriteria class intervals for choropleth maps". Annals, Association of American Geographers, 93 (3), 595-623

Jenks, G. F., Caspall, F. C., 1971. "Error on choroplethic maps: definition, measurement, reduction". Annals, Association of American Geographers, 61 (2), 217-244

Dent, B. D., 1999, Cartography: thematic map design. McGraw-Hill, Boston, 417 pp. Slocum TA, McMaster RB, Kessler FC, Howard HH 2005 Thematic Cartography and Geographic Visualization, Prentice Hall, Upper Saddle River NJ.

Fisher, W. D. 1958 "On grouping for maximum homogeneity", Journal of the American Statistical Association, 53, 789-798

See also

Examples

landscape_classified <- classify_habitats(landscape, classes = 5)