energy_fun

Energy function

Usage

Energy_fun(
  f,
  f0,
  statistics,
  f_,
  f0_,
  statistics_,
  fn,
  p,
  p_,
  Lp,
  w_statistics
)

Arguments

f

Result of the calc_moments_full function which represents product-moment contribution of a point at coordinates x, y with marks, for the whole new ponit pattern.

f0

Column sums of the weights of the brand correlation functions of the new point pattern.

statistics

Results of the compute_statistics function for the new point pattern (calculation of optional spatial statistics).

fn

Determination of the weightings of the mark correlation functions.

p

Defines the initial state of the new ponit pattern.

p_

Reference point pattern.

Lp

Distance measure for the calculation of the energy function (Lp distance, 1 <= p <Inf).

w_statistics

Vector of named weights for optional spatial statistics from the spatstat package to be included in the energy calculation.This may include Dk, K, Hs, pcf.

Value

vector

Details

Defining the Energy_fun function to calculate the "energy" of the pattern (where a lower energy indicates a better match).