fit_point_process

Fit point process to randomize data

fit_point_process(
  pattern,
  n_random = 1,
  process = "poisson",
  return_input = TRUE,
  simplify = FALSE,
  verbose = TRUE
)

Arguments

pattern
ppp object with point pattern

n_random
Integer with number of randomizations.

process
Character specifying which point process model to use. Either "poisson" or "cluster".

return_input
Logical if the original input data is returned.

simplify
Logical if only pattern will be returned if n_random = 1 and return_input = FALSE.

verbose
Logical if progress report is printed.

Value

rd_pat

Details

The functions randomizes the observed point pattern by fitting a point process to the data and simulating n_random patterns using the fitted point process. It is possible to choose between a Poisson process or a Thomas cluster process model. For more information about the point process models, see e.g. Wiegand & Moloney (2014).

References

Plotkin, J. B., Potts, M. D., Leslie, N., Manokaran, N., LaFrankie, J. V., & Ashton, P. S. (2000). Species-area curves, spatial aggregation, and habitat specialization in tropical forests. Journal of Theoretical Biology, 207(1), 81-99.

Wiegand, T., Moloney, K.A., 2014. Handbook of spatial point-pattern analysis in ecology. Chapman and Hall/CRC Press, Boca Raton.

Examples

pattern_fitted <- fit_point_process(pattern = species_a, n_random = 39)

#> Warning: Unmarked provided input pattern.

#> > Progress: n_random: 1/39

#> > Progress: n_random: 2/39

#> > Progress: n_random: 3/39

#> > Progress: n_random: 4/39

#> > Progress: n_random: 5/39

#> > Progress: n_random: 6/39

#> > Progress: n_random: 7/39

#> > Progress: n_random: 8/39

#> > Progress: n_random: 9/39

#> > Progress: n_random: 10/39

#> > Progress: n_random: 11/39

#> > Progress: n_random: 12/39

#> > Progress: n_random: 13/39

#> > Progress: n_random: 14/39

#> > Progress: n_random: 15/39

#> > Progress: n_random: 16/39

#> > Progress: n_random: 17/39

#> > Progress: n_random: 18/39

#> > Progress: n_random: 19/39

#> > Progress: n_random: 20/39

#> > Progress: n_random: 21/39

#> > Progress: n_random: 22/39

#> > Progress: n_random: 23/39

#> > Progress: n_random: 24/39

#> > Progress: n_random: 25/39

#> > Progress: n_random: 26/39

#> > Progress: n_random: 27/39

#> > Progress: n_random: 28/39

#> > Progress: n_random: 29/39

#> > Progress: n_random: 30/39

#> > Progress: n_random: 31/39

#> > Progress: n_random: 32/39

#> > Progress: n_random: 33/39

#> > Progress: n_random: 34/39

#> > Progress: n_random: 35/39

#> > Progress: n_random: 36/39

#> > Progress: n_random: 37/39

#> > Progress: n_random: 38/39

#> > Progress: n_random: 39/39

#>