plot.rd_mar

Plot method for rd_pat object

Usage

# S3 method for class 'rd_mar'
plot(
  x,
  what = "sf",
  n = NULL,
  probs = c(0.025, 0.975),
  ask = TRUE,
  verbose = TRUE,
  ...
)

Arguments

x

rd_mar object with randomized patterns.

what

Character specifying to plot summary functions of point patterns (what = "sf") or actual patterns (what = "pp").

n

Integer with number or vector of ids of randomized pattern to plot. See Details section for more information.

probs

Vector with quantiles of randomized data used for envelope construction.

ask

Logical if the user is asked to press <RETURN> before second summary function is plotted (only used if what = "sf").

verbose

Logical if progress report is printed.

...

Not used.

Value

void

Details

The function plots the pair correlation function and the nearest neighbour function of the observed pattern and the reconstructed patterns (as "simulation envelopes").

It is also possible to plot n randomized patterns and the observed pattern using what = "pp". If n is a single number, n randomized patterns will be sampled to plot. If n is a vector, the corresponding patterns will be plotted.

See also

Examples

if (FALSE) { # \dontrun{
pattern_recon <- reconstruct_pattern(species_a, n_random = 1, max_runs = 1000,
simplify = TRUE, return_input = FALSE)
marks_sub <- spatstat.geom::subset.ppp(species_a, select = dbh)
marks_recon <- reconstruct_pattern_marks(pattern_recon, marks_sub,
n_random = 19, max_runs = 1000)
plot(marks_recon)
} # }