plot_randomized_raster

Plot randomized raster

plot_randomized_raster(raster, n = NULL, col, verbose = TRUE, nrow, ncol)

Arguments

raster
rd_ras object with randomized raster.

n
Integer with number or vector of ids of randomized raster to plot. See Details section for more information.

col
Vector with color palette used for plotting.

verbose
Logical if messages are printed.

nrow, ncol
Integer with number of rows and columns of plot grid.

Value

void

Details

Function to plot randomized raster. If n is a single number, n randomized raster will be sampled to plot. If n is a vector, the corresponding raster will be plotted. col, nrow, ncol are passed to plot.

See also

Examples

if (FALSE) {
landscape_classes <- classify_habitats(raster = landscape, classes = 5)
landscape_random <- randomize_raster(raster = landscape_classes, n_random = 19)

plot_randomized_raster(landscape_random)

palette <- viridis::viridis(n = 5)
plot_randomized_raster(landscape_random, n = 5, col = palette, nrow = 3, ncol = 2)
}