reconstruct_algorithm

Pattern reconstruction (internal)

Usage

reconstruct_algorithm(
  pattern,
  method,
  n_random,
  e_threshold,
  max_runs,
  no_change,
  annealing,
  weights,
  r_length,
  r_max,
  stoyan,
  verbose,
  plot
)

Arguments

pattern

ppp object with pattern.

n_random

Integer with number of randomizations.

e_threshold

Double with minimum energy to stop reconstruction.

max_runs

Integer with maximum number of iterations if e_threshold is not reached.

no_change

Integer with number of iterations at which the reconstruction will stop if the energy does not decrease.

annealing

Double with probability to keep relocated point even if energy did not decrease.

weights

Vector with weights used to calculate energy. The first number refers to Gest(r), the second number to pcf(r).

r_length

Integer with number of intervals from r = 0 to r = rmax for which the summary functions are evaluated.

r_max

Double with maximum distance used during calculation of summary functions. If NULL, will be estimated from data.

stoyan

Coefficient for Stoyan's bandwidth selection rule.

verbose

Logical if progress report is printed.

plot

Logical if pcf(r) function is plotted and updated during optimization.

Value

list