reconstruct_pattern_hetero

Pattern reconstruction for heterogeneous patterns

Usage

reconstruct_pattern_hetero(
  pattern,
  n_random = 1,
  e_threshold = 0.01,
  max_runs = 1000,
  no_change = Inf,
  annealing = 0.01,
  comp_fast = 1000,
  weights = c(0.5, 0.5),
  r_length = 250,
  return_input = TRUE,
  simplify = FALSE,
  verbose = TRUE,
  plot = FALSE
)

Arguments

pattern

ppp object with pattern.

n_random

Integer with number of randomizations.

e_threshold

Double with minimum energy to stop reconstruction.

max_runs

Integer with maximum number of iterations if e_threshold is not reached.

no_change

Integer with number of iterations at which the reconstruction will stop if the energy does not decrease.

annealing

Double with probability to keep relocated point even if energy did not decrease.

comp_fast

Integer with threshold at which summary functions are estimated in a computational fast way.

weights

Vector with weights used to calculate energy. The first number refers to Gest(r), the second number to pcf(r).

r_length

Integer with number of intervals from r = 0 to r = rmax for which the summary functions are evaluated.

return_input

Logical if the original input data is returned.

simplify

Logical if only pattern will be returned if n_random = 1 and return_input = FALSE.

verbose

Logical if progress report is printed.

plot

Logical if pcf(r) function is plotted and updated during optimization.

Value

rd_pat

References

Kirkpatrick, S., Gelatt, C.D.Jr., Vecchi, M.P., 1983. Optimization by simulated annealing. Science 220, 671–680. <https://doi.org/10.1126/science.220.4598.671>

Tscheschel, A., Stoyan, D., 2006. Statistical reconstruction of random point patterns. Computational Statistics and Data Analysis 51, 859–871. <https://doi.org/10.1016/j.csda.2005.09.007>

Wiegand, T., Moloney, K.A., 2014. Handbook of spatial point-pattern analysis in ecology. Chapman and Hall/CRC Press, Boca Raton. ISBN 978-1-4200-8254-8

Examples

if (FALSE) {
input_pattern <- spatstat.random::rpoispp(lambda = function(x, y) {100 * exp(-3 * x)},
nsim = 1)
pattern_recon <- reconstruct_pattern_hetero(input_pattern, n_random = 19, max_runs = 1000)
}