Package description

Analyse species-habitat associations in R. Therefore, information about the location of the species is needed and about the environmental conditions. To test for significance habitat associations, one of the two components is randomized. Methods are mainly based on Plotkin et al. (2000) <doi:10.1006/jtbi.2000.2158> and Harms et al. (2001) <doi:10.1111/j.1365-2745.2001.00615.x>.

See also

Author

Maintainer: Maximilian H.K. Hesselbarth mhk.hesselbarth@gmail.com (ORCID)

Authors:

Marco Sciaini marco.sciaini@posteo.net (ORCID)

Chris Wudel chris.wudel@tu-dresden.de (ORCID)

Other contributors:

Zeke Marshall ee18zm@leeds.ac.uk (ORCID) [contributor]

Thomas Etherington teth001@aucklanduni.ac.nz (ORCID) [contributor]

Janosch Heinermann janosch.heinermann@tu-dresden.de (ORCID) [contributor]