calculate_energy

Calculate mean energy

calculate_energy(pattern, return_mean = FALSE, comp_fast = FALSE,
  verbose = TRUE)

Arguments

pattern
List with reconstructed patterns.

return_mean
Return the mean energy

comp_fast
Should summary functions be estimated in an computational fast way.

verbose
Print progress report.

Value

numeric

Details

The function calculates the mean energy (or deviation) between the observed pattern and all reconstructed patterns (for more information see Tscheschel & Stoyan (2006) or Wiegand & Moloney (2014)). The pair correlation function and the nearest neighbour distance function are used to describe the patterns. For large patterns `comp_fast = TRUE` decreases the computational demand because no edge correction is used and the pair correlation function is estimated based on Ripley's K-function. For more information see estimate_pcf_fast.

References

Tscheschel, A., & Stoyan, D. (2006). Statistical reconstruction of random point patterns. Computational Statistics and Data Analysis, 51(2), 859–871.

Wiegand, T., & Moloney, K. A. (2014). Handbook of spatial point-pattern analysis in ecology. Boca Raton: Chapman and Hall/CRC Press.

See also

Examples

pattern_random <- fit_point_process(species_a, n_random = 39)

#> > Progress: n_random: 1/39

#> > Progress: n_random: 2/39

#> > Progress: n_random: 3/39

#> > Progress: n_random: 4/39

#> > Progress: n_random: 5/39

#> > Progress: n_random: 6/39

#> > Progress: n_random: 7/39

#> > Progress: n_random: 8/39

#> > Progress: n_random: 9/39

#> > Progress: n_random: 10/39

#> > Progress: n_random: 11/39

#> > Progress: n_random: 12/39

#> > Progress: n_random: 13/39

#> > Progress: n_random: 14/39

#> > Progress: n_random: 15/39

#> > Progress: n_random: 16/39

#> > Progress: n_random: 17/39

#> > Progress: n_random: 18/39

#> > Progress: n_random: 19/39

#> > Progress: n_random: 20/39

#> > Progress: n_random: 21/39

#> > Progress: n_random: 22/39

#> > Progress: n_random: 23/39

#> > Progress: n_random: 24/39

#> > Progress: n_random: 25/39

#> > Progress: n_random: 26/39

#> > Progress: n_random: 27/39

#> > Progress: n_random: 28/39

#> > Progress: n_random: 29/39

#> > Progress: n_random: 30/39

#> > Progress: n_random: 31/39

#> > Progress: n_random: 32/39

#> > Progress: n_random: 33/39

#> > Progress: n_random: 34/39

#> > Progress: n_random: 35/39

#> > Progress: n_random: 36/39

#> > Progress: n_random: 37/39

#> > Progress: n_random: 38/39

#> > Progress: n_random: 39/39

#> 

calculate_energy(pattern_random)

#> > Progress: 1/39

#> > Progress: 2/39

#> > Progress: 3/39

#> > Progress: 4/39

#> > Progress: 5/39

#> > Progress: 6/39

#> > Progress: 7/39

#> > Progress: 8/39

#> > Progress: 9/39

#> > Progress: 10/39

#> > Progress: 11/39

#> > Progress: 12/39

#> > Progress: 13/39

#> > Progress: 14/39

#> > Progress: 15/39

#> > Progress: 16/39

#> > Progress: 17/39

#> > Progress: 18/39

#> > Progress: 19/39

#> > Progress: 20/39

#> > Progress: 21/39

#> > Progress: 22/39

#> > Progress: 23/39

#> > Progress: 24/39

#> > Progress: 25/39

#> > Progress: 26/39

#> > Progress: 27/39

#> > Progress: 28/39

#> > Progress: 29/39

#> > Progress: 30/39

#> > Progress: 31/39

#> > Progress: 32/39

#> > Progress: 33/39

#> > Progress: 34/39

#> > Progress: 35/39

#> > Progress: 36/39

#> > Progress: 37/39

#> > Progress: 38/39

#> > Progress: 39/39

#> 

#>  [1] 0.4066292 0.3788558 0.4915708 0.3605971 0.3999073 0.3176181 0.3358566
#>  [8] 0.3919977 0.3857408 0.4138439 0.4032968 0.3677975 0.3335413 0.3325191
#> [15] 0.4517818 0.3892915 0.3937594 0.2829634 0.3962145 0.3427752 0.3804610
#> [22] 0.4646247 0.3486880 0.3813034 0.3239423 0.3532702 0.4691708 0.4104261
#> [29] 0.3507668 0.2732523 0.4029452 0.4248271 0.4016035 0.4291119 0.3904745
#> [36] 0.3258776 0.4654834 0.4045895 0.4698011

calculate_energy(pattern_random, return_mean = TRUE)

#> > Progress: 1/39

#> > Progress: 2/39

#> > Progress: 3/39

#> > Progress: 4/39

#> > Progress: 5/39

#> > Progress: 6/39

#> > Progress: 7/39

#> > Progress: 8/39

#> > Progress: 9/39

#> > Progress: 10/39

#> > Progress: 11/39

#> > Progress: 12/39

#> > Progress: 13/39

#> > Progress: 14/39

#> > Progress: 15/39

#> > Progress: 16/39

#> > Progress: 17/39

#> > Progress: 18/39

#> > Progress: 19/39

#> > Progress: 20/39

#> > Progress: 21/39

#> > Progress: 22/39

#> > Progress: 23/39

#> > Progress: 24/39

#> > Progress: 25/39

#> > Progress: 26/39

#> > Progress: 27/39

#> > Progress: 28/39

#> > Progress: 29/39

#> > Progress: 30/39

#> > Progress: 31/39

#> > Progress: 32/39

#> > Progress: 33/39

#> > Progress: 34/39

#> > Progress: 35/39

#> > Progress: 36/39

#> > Progress: 37/39

#> > Progress: 38/39

#> > Progress: 39/39

#> 

#> [1] 0.3858251