reconstruct_pattern_hetero

Pattern reconstruction for heterogeneous patterns

reconstruct_pattern_hetero(
  pattern,
  n_random = 1,
  e_threshold = 0.01,
  max_runs = 1000,
  no_change = Inf,
  annealing = 0.01,
  comp_fast = 1000,
  weights = c(0.5, 0.5),
  r_length = 250,
  return_input = TRUE,
  simplify = FALSE,
  verbose = TRUE,
  plot = FALSE
)

Arguments

pattern
ppp object with pattern.

n_random
Integer with number of randomizations.

e_threshold
Double with minimum energy to stop reconstruction.

max_runs
Integer with maximum number of iterations if e_threshold is not reached.

no_change
Integer with number of iterations at which the reconstruction will stop if the energy does not decrease.

annealing
Double with probability to keep relocated point even if energy did not decrease.

comp_fast
Integer with threshold at which summary functions are estimated in a computational fast way.

weights
Vector with weights used to calculate energy. The first number refers to Gest(r), the second number to pcf(r).

r_length
Integer with number of intervals from r = 0 to r = rmax for which the summary functions are evaluated.

return_input
Logical if the original input data is returned.

simplify
Logical if only pattern will be returned if n_random = 1 and return_input = FALSE.

verbose
Logical if progress report is printed.

plot
Logical if pcf(r) function is plotted and updated during optimization.

Value

rd_pat

Details

The functions randomizes the observed pattern by using pattern reconstruction as described in Tscheschel & Stoyan (2006) and Wiegand & Moloney (2014). The algorithm starts with a random but heterogeneous pattern, shifts a point to a new location and keeps the change only, if the deviation between the observed and the reconstructed pattern decreases. The pair correlation function and the nearest neighbour distance function are used to describe the patterns.

For large patterns (n > comp_fast) the pair correlation function can be estimated from Ripley's K-function without edge correction. This decreases the computational time. For more information see estimate_pcf_fast.

The reconstruction can be stopped automatically if for n steps the energy does not decrease. The number of steps can be controlled by no_change and is set to no_change = Inf as default to never stop automatically.

The weights must be 0 < sum(weights) <= 1. To weight both summary functions identical, use weights = c(0.5, 0.5).

spatstat sets r_length to 513 by default. However, a lower value decreases the computational time while increasing the "bumpiness" of the summary function.

References

Tscheschel, A., & Stoyan, D. (2006). Statistical reconstruction of random point patterns. Computational Statistics and Data Analysis, 51(2), 859-871.

Wiegand, T., & Moloney, K. A. (2014). Handbook of spatial point-pattern analysis in ecology. Boca Raton: Chapman and Hall/CRC Press.

See also

Examples

if (FALSE) {
input_pattern <- spatstat.core::rpoispp(lambda = function(x, y) {100 * exp(-3 * x)}, nsim = 1)

pattern_recon <- reconstruct_pattern_hetero(input_pattern, n_random = 19, max_runs = 1000)
}